feat: parallel batch [T20260509-3]
Tasks:
- T20260509-3: Modularize crates/orbit-engine/src/activity_job/cli_runner.rs (2,161 lines)
